Added configuration to the whatsapp module
