+++ lib/whatsapp_sdk.rb
@@ -1,38 +1,17 @@
 # frozen_string_literal: true
 
-# Configuration
-require_relative "whatsapp_sdk/configuration"
+require "zeitwerk"
+loader = Zeitwerk::Loader.for_gem
+loader.setup # ready!
 
-# APIs
-require_relative "whatsapp_sdk/api/phone_numbers"
-require_relative "whatsapp_sdk/api/messages"
-require_relative "whatsapp_sdk/api/medias"
-require_relative "whatsapp_sdk/api/client"
+module WhatsappSdk
+  class << self
+    def configuration
+      @configuration ||= Configuration.new
+    end
 
-# APIs responses
-require_relative "whatsapp_sdk/api/responses/message_data_response"
-require_relative "whatsapp_sdk/api/responses/phone_number_data_response"
-require_relative "whatsapp_sdk/api/responses/phone_numbers_data_response"
-require_relative "whatsapp_sdk/api/responses/message_error_response"
-require_relative "whatsapp_sdk/api/responses/data_response"
-require_relative "whatsapp_sdk/api/responses/read_message_data_response"
-require_relative "whatsapp_sdk/api/responses/media_data_response"
-require_relative "whatsapp_sdk/api/responses/success_response"
-require_relative "whatsapp_sdk/api/responses/error_response"
-
-# Resources
-require_relative "whatsapp_sdk/resource/address"
-require_relative "whatsapp_sdk/resource/button_parameter"
-require_relative "whatsapp_sdk/resource/component"
-require_relative "whatsapp_sdk/resource/contact_response"
-require_relative "whatsapp_sdk/resource/contact"
-require_relative "whatsapp_sdk/resource/currency"
-require_relative "whatsapp_sdk/resource/date_time"
-require_relative "whatsapp_sdk/resource/email"
-require_relative "whatsapp_sdk/resource/media"
-require_relative "whatsapp_sdk/resource/message"
-require_relative "whatsapp_sdk/resource/name"
-require_relative "whatsapp_sdk/resource/org"
-require_relative "whatsapp_sdk/resource/parameter_object"
-require_relative "whatsapp_sdk/resource/phone_number"
-require_relative "whatsapp_sdk/resource/url"
+    def configure
+      yield(configuration)
+    end
+  end
+end
